Refactor Grace Boot
* Move `ApplicationArtefactHandler` `ApplicationClassInjector` from grace-core to grace-boot
* Move `LoggingTransformerSpec` from grace-logging to grace-boot
* Update `ArtefactTypeAstTransformationSpec` `GlobalGrailsClassInjectorTransformationSpec` `GrailsAwareInjectionOperationSpec`
